Working to fix IOCP
Also added the DNS module. As of now it provides a method for
asynchronous Linux addr lookups. Working to make Linux work properly and
have the same on windows. Linux uses `getaddrinfo_a` and Windows will
use `GetAddrInfoEx` for the most part. Both require a little bit of
ctypes which isnt my strength